[#11] feat: 현재 경도, 위도를 추적하여 미세먼지 수치 제공

+++ AirPollution-Weather-APP/AirPollution-Weather-APP/Model/AirPollutionApiModel.swift
@@ -15,8 +15,8 @@ struct AirPollutionDatasStruct: Codable {
 
 // MARK: - Coord
 struct Coord: Codable {
-    let lon: Int
-    let lat: Int
+    let lon: Double
+    let lat: Double
 }
 
 // MARK: - List
